\name{ellipsoidNormal}
\title{Ellipsoid normal vector}
\alias{ellipsoidNormal}
\description{Normal vector of an ellipsoid at given point in cartesian
  coordinates. }
\usage{ellipsoidNormal( latlon )}
\arguments{
  \item{latlon}{ c( latitude [deg] , longitude [deg] ) }
}
\value{

  c(x,y,z), normal vector in geocentric cartesian coordinates. 
  
}

\author{Ilkka Virtanen (University of Oulu, Finland) \cr
  \email{ilkka.i.virtanen@oulu.fi}}