\name{radarSites}
\title{IS radar site coordinates}
\alias{radarSites}
\description{Coordinates of incoherent scatter radars }
\usage{radarSites()}
\arguments{None}
\value{ A list of radar site coordinates. Each element is a vector of
  the form c(latitude [deg], longitude [deg], height [m]) }

\author{Ilkka Virtanen (University of Oulu, Finland) \cr
  \email{ilkka.i.virtanen@oulu.fi}}